Refactor curator app
